accept variations on relatinship names; warn don't crash if cannot process.
